Updated demos
Mug cleanup has two new environments. One that spawns a cube on the table and another that spawns a cube in a mug on the table.

Nut assembly now starts with both nuts on the cylinder peg.

Demo random action has some code to change the camera placement and orientation.

requirements.txt has been updated to remove 2 redundant packages.